backend/requirements.txt
Flask==3.0.0
flask-cors
kafka-python
six
python-dotenv
requests